AI/requirements.txt
absl-py==2.1.0
adapters==0.2.1
aiofiles==23.1.0
aiohttp==3.7.4.post0
aiosignal==1.3.1
alkana==0.0.3
anyio==4.4.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==3.0.1
attrs==23.2.0
audioread==3.0.1
beautifulsoup4==4.12.2
blinker==1.8.2
Brotli==1.0.9
cached-property==1.5.2
certifi==2024.7.4
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.5
decorator==5.1.1
discord.py==2.4.0
docker==7.1.0
docopt==0.6.2
elevenlabslib==0.22.5
emoji==2.12.1
exceptiongroup==1.2.2
ffmpeg==1.4
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
gast==0.6.0
gevent==22.10.2
google-pasta==0.2.0
googletrans==4.0.0rc1
greenlet==2.0.2
grpcio==1.64.1
h11==0.9.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.24.2
humanfriendly==10.0
humanize==4.6.0
hyperframe==5.2.0
idna==2.10
inflate64==0.3.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Js2Py==0.74
keras==3.4.1
keyboard==0.13.5
lazy_loader==0.4
Levenshtein==0.25.1
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mecab==0.996.3
mecab-python3==1.0.9
mkl==2021.4.0
ml-dtypes==0.4.0
mosestokenizer==1.2.1
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.2
multivolumefile==0.2.3
mutagen==1.47.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.7.0
numba==0.60.0
numpy==1.26.4
onnxruntime==1.18.1
openai==0.28.1
openfile==0.0.7
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
Pillow==9.3.0
pipwin==0.5.2
plac==1.4.3
platformdirs==4.2.2
pooch==1.8.2
protobuf==4.25.3
psutil==5.9.5
py7zr==0.20.5
PyAudio==0.2.14
PyAutoGUI==0.9.54
pybcj==1.0.1
pycparser==2.20
pycryptodomex==3.17
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyjsparser==2.7.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyperclip==1.9.0
pypiwin32==223
pyppmd==1.0.0
PyPrind==2.11.3
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
pySmartDL==1.3.4
PySocks==1.7.1
pytchat==0.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pyttsx3==2.90
pytube==15.0.0
pytweening==1.2.0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pyvts==0.3.0
pywin32==306
PyYAML==6.0.1
pyzstd==0.15.6
rapidfuzz==3.9.4
reg==0.12
regex==2024.7.24
repoze.lru==0.7
requests==2.32.3
resampy==0.4.3
rfc3986==1.5.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.23.1
six==1.16.0
skops==0.10.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.4.0
sympy==1.13.0
tabulate==0.9.0
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
texttable==1.6.7
threadpoolctl==3.5.0
tokenizers==0.15.2
toolwrapper==2.1.0
torch==2.3.1
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.4
transformers==4.39.3
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==4.3
uctools==1.3.0
unidic==1.1.0
unidic-lite==1.0.8
urllib3==2.2.2
vboxapi==1.0
wasabi==0.10.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
wtpsplit==2.0.5
xmltodict==0.13.0
yarl==1.9.4
youtube-dl==2021.12.17
youtube-search-python==1.6.6
yt-dlp==2024.8.6
zope.event==4.6
zope.interface==6.0
